+++ app/gui/src/dashboard/components/ProfilePicture/index.ts
@@ -1,1 +1,2 @@
 export * from './ProfilePicture'
+export * from './variants'
